Try to make it even easier to read.
